requirements.txt
Django==1.5
argparse==1.2.1
distribute==0.6.24
dj-database-url==0.2.1
django-tastypie==0.9.12
gunicorn==0.17.2
ipython==0.13.1
mimeparse==0.1.3
pusher==0.7
pycrypto==2.6
python-dateutil==2.1
requests==2.0.1
six==1.2.0
wsgiref==0.1.2
